requirements.txt
langchain==0.1.15
openai==1.16.2
python-dotenv==1.0.1
googlesearch-python==1.2.3
langchain_openai==0.1.1
langchainhub
selenium==4.19.0
webdriver-manager==4.0.1
pydantic-settings==2.0.3
tweepy==4.14.0